feat(math): migrate GEMV dispatchers and A2 weights from f16/u16 to native f32
- Remove duplicate rechannel_w field from WaveNetA2<CH> (SQ4.6)
- Change SimdMath trait + 4 impls fused_add_gemv/gemv_overwrite/gemv_overwrite_batch from &[u16] to &[f32] (SQ4.7)
- Update scalar fallbacks (scalar_ref/gemm.rs, scalar_ref/lstm.rs) — remove f16_bits_to_f32, use f32 directly
- Fix tests/benches: lstm_test, soak_test, lstm_scalar_bf16_parity, lstm_gate_bf16_parity, gemv_bench
- Delete dead gemv_4gate_fallback (no callers)
- cargo check, cargo test --lib, cargo clippy: all clean (31 pre-existing warnings only)

+++ src/math/common/avx2_impl.rs
@@ -238,7 +238,7 @@ impl SimdMath for Avx2Math {
     // SAFETY: slices are valid; CPU supports AVX2+FMA (x86-64-v3, verified by dispatch).
     unsafe fn fused_add_gemv(
         in_frame: &[f32],
-        weights: &[u16],
+        weights: &[f32],
         bias: &[f32],
         out_frame: &mut [f32],
         do_bias: bool,
@@ -320,7 +320,7 @@ impl SimdMath for Avx2Math {
     // SAFETY: slices are valid; CPU supports AVX2+FMA (x86-64-v3, verified by dispatch).
     unsafe fn gemv_overwrite(
         in_frame: &[f32],
-        weights: &[u16],
+        weights: &[f32],
         bias: &[f32],
         out_frame: &mut [f32],
         do_bias: bool,
@@ -929,7 +929,7 @@ impl SimdMath for Avx2Math {
     // SAFETY: slices are valid; CPU supports AVX2+FMA (x86-64-v3, verified by dispatch).
     unsafe fn gemv_overwrite_batch(
         in_frames: &[f32],
-        weights: &[u16],
+        weights: &[f32],
         bias: &[f32],
         out_frames: &mut [f32],
         num_frames: usize,
