+++ src/math/common/avx512/gemv/vnni_bf16.rs
@@ -164,7 +164,7 @@ macro_rules! impl_avx512vnni_bf16_gemv {
         // CPU supports AVX-512 VNNI+BF16.
         unsafe fn fused_add_gemv(
             in_frame: &[f32],
-            weights: &[u16],
+            weights: &[f32],
             bias: &[f32],
             out_frame: &mut [f32],
             do_bias: bool,
@@ -223,7 +223,7 @@ macro_rules! impl_avx512vnni_bf16_gemv {
         // CPU supports AVX-512 VNNI+BF16.
         unsafe fn gemv_overwrite(
             in_frame: &[f32],
-            weights: &[u16],
+            weights: &[f32],
             bias: &[f32],
             out_frame: &mut [f32],
             do_bias: bool,
@@ -288,7 +288,7 @@ macro_rules! impl_avx512vnni_bf16_gemv {
         // CPU supports AVX-512 VNNI+BF16.
         unsafe fn gemv_overwrite_batch(
             in_frames: &[f32],
-            weights: &[u16],
+            weights: &[f32],
             bias: &[f32],
             out_frames: &mut [f32],
             num_frames: usize,
